contrib/marconi-plugin/requirements.txt
python-marconiclient>=0.0.1a1
